Moved color into pretty
